add nodemailer

+++ src/controllers/vuelosController.js
@@ -1,6 +1,16 @@
 import VuelosDao from "../dao/vuelosDAO.js";
 const vuelosDao = new VuelosDao();
 
+import UsersDao from "../dao/usersDao.js";
+const userDao = new UsersDao();
+
+import nodemailer from 'nodemailer';
+import dotenv from 'dotenv';
+
+dotenv.config();
+const EMAIL = process.env.EMAIL_USER;
+const PASS = process.env.EMAIL_PASS;
+
 class VuelosManager {
 
     async getAllVuelos() {
@@ -48,7 +58,7 @@ class VuelosManager {
     }
 
     async createVuelos(vuelo) {
-        const { empresa, origen, destino, vuelo_ida, vuelo_vuelta, precio, duracion, clase, asientos_disponibles, incluye_equipaje, pasajeros } = vuelo;
+        const { empresa, origen, destino, vuelo_ida, vuelo_vuelta, precio, duracion, clase, asientos_disponibles, incluye_equipaje, pasajeros, owner } = vuelo;
 
         if (!empresa || !origen || !destino || !vuelo_ida || !duracion || !clase || typeof precio !== 'number' || typeof asientos_disponibles !== 'number' ||
             !Array.isArray(pasajeros)
@@ -56,7 +66,7 @@ class VuelosManager {
             throw new Error("Faltan campos obligatorios o tienen el tipo incorrecto");
         }
         try {
-            const result = await vuelosDao.createVuelosDao({ empresa, origen, destino, vuelo_ida, vuelo_vuelta, precio, duracion, clase, asientos_disponibles, incluye_equipaje, pasajeros })
+            const result = await vuelosDao.createVuelosDao({ empresa, origen, destino, vuelo_ida, vuelo_vuelta, precio, duracion, clase, asientos_disponibles, incluye_equipaje, pasajeros, owner })
             return result;
         } catch (error) {
             throw new Error(`Error al crear los vuelos ${error.message}`);
@@ -83,11 +93,46 @@ class VuelosManager {
         const vuelo = await this.getVuelosById(vid);
         if (!vuelo) throw new Error(`El vuelo con ${vid} no se encuentra`);
         try {
-            const result = await vuelosDao.deleteVuelosDao(vid);
-            return result;
+
+            const ownerEmail = vuelo.owner;
+            if (ownerEmail === "admin")
+                return await vuelosDao.deleteVuelosDao(vid);
+
+            const user = await userDao.getEmailDao(ownerEmail);
+            if (user?.role === 'premium') {
+                this.sendEmailProductDelete(ownerEmail, vid);
+
+                const result = await vuelosDao.deleteVuelosDao(vid);
+                return result;
+            }
         } catch (error) {
             throw new Error(`Error al actualizar el vuelo ${error.message}`);
         }
+    }
+
+    async sendEmailProductDelete(email, productId) {
+        const transport = nodemailer.createTransport({
+            service: 'gmail',
+            port: 587,
+            auth: {
+                user: EMAIL,
+                pass: PASS
+            }
+        });
+
+        await transport.sendMail({
+            from: 'Edgar Steinberg <s.steinberg2019@gmail.com>',
+            to: email,
+            subject: 'Eliminación de Producto',
+            html: `<div style="font-family: Arial, sans-serif; color: #333;">
+                                        <h1>Notificación de Eliminación de Producto</h1>
+                                        <p>El producto con ID ${productId} ha sido eliminado de la plataforma.</p>
+                                        <p>Si tienes alguna pregunta, por favor contáctanos.</p>
+                                        <p>Gracias,</p>
+                                        <p>El equipo de soporte </p>
+                                        </div>`,
+        });
+
     }
 }
